Move Job class to separate file.
