+++ src/Core/FSpot.Database/FSpot.Database.csproj
@@ -100,6 +100,7 @@
     <Compile Include="PhotoEventArgs.cs" />
     <Compile Include="IUpdaterUI.cs" />
     <Compile Include="Jobs\CalculateHashJob.cs" />
+    <Compile Include="Jobs\Job.cs" />
     <Compile Include="Jobs\SyncMetadataJob.cs" />
     <Compile Include="IDb.cs" />
   </ItemGroup>
